requirements.txt
yfinance
pandas
numpy
scikit-learn
tensorflow
matplotlib
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
beautifulsoup4
nltk
python-dotenv
